Portfolio updated so each script produces a heuristic value. Added progressive unpruning.

+++ src/players/portfolio/scripts/utils/ValuePoint.java
@@ -5,5 +5,5 @@ import core.game.GameState;
 
 public interface ValuePoint
 {
-    int ofInterest(GameState gs, Actor ac, int posX, int posY);
+    double ofInterest(GameState gs, Actor ac, int posX, int posY);
 }
